pyproject.toml
[tool.poetry]
name = "i3-meetings"
version = "0.1.0"
description = ""
authors = ["polydome <polydome@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytz = "^2021.1"
google-api-python-client = "^2.2.0"
py3status = "^3.36"
google-auth = "^1.29.0"
google-auth-oauthlib = "^0.4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
